+++ MelonLoader/Attributes/HideFromIl2CppHarmonyPatcher.cs
@@ -1,0 +1,7 @@
+﻿using System;
+
+namespace MelonLoader.Attributes
+{
+    [AttributeUsage(AttributeTargets.Method | AttributeTargets.Class)]
+    public class HideFromIl2CppHarmonyPatcherAttribute : Attribute { public HideFromIl2CppHarmonyPatcherAttribute() { } }
+}