+++ CHANGELOG.md
@@ -81,6 +81,7 @@
 48. Fixed Issue with Solution Build Dependencies copying to Output Directory.
 49. Fixed Issue with Il2Cpp Harmony Patcher breaking on Custom Il2Cpp Types.   (Credits to [ds5678](https://github.com/ds5678) :D)
 50. Implemented Management and Disposal Handling for Start Screen UI Elements.
+51. Added HideFromIl2CppHarmonyPatcher Attribute.
 
 ---